+++ src/github.com/getlantern/flashlight/null_systray.go
@@ -1,0 +1,15 @@
+// +build headless
+
+package main
+
+func runOnSystrayReady(f func()) {
+	showui = false
+	f()
+}
+
+func quitSystray() {
+}
+
+func configureSystemTray() error {
+	return nil
+}
